requirements.txt
uplink
requests
marshmallow
python-dotenv